Created newCustomer(),
this contains 5 new methods. frameSize(), which was created due to code
duplication, this method is called numerous amount of times throughout
the program, same with the windowlistener() and addPanelInfo().

Due to newCustomer(), being such a large method, I decided to break it
down when it came to the add and cancel buttons, created 2 new methods
to be called, which makes the code a lot easier to understand when you
read it